% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseBCF.R
\name{predict_SparseBCF}
\alias{predict_SparseBCF}
\title{Takes a fitted SparseBCF object and produces predictions for a new set of covariate values}
\usage{
predict_SparseBCF(
  model,
  x_predict_moderate,
  x_predict_control = x_predict_moderate,
  pi_pred = rep(1, nrow(x_predict_control)),
  type = "tau"
)
}
\arguments{
\item{model}{output from SparseBCF}

\item{x_predict_moderate}{matrix of covariates for the covariate-dependent treatment effects tau(x) for predictions (optional)}

\item{x_predict_control}{matrix of covariates for the "prognostic" function mu(x) for predictions (optional)}

\item{pi_pred}{propensity score for prediction}

\item{type}{either "tau" or "mu" to indicate which type of prediction should be made}
}
\description{
This function requires that you indicate where the trees from the model fit are saved. By default, 
You can do so by specifying the save_tree_directory argument in SparseBCF().
}
